recreating dataset page as a dataset + summary statistics page. similar layout planned as that for relationship page
